+++ api/violation/deleteViolation.php
@@ -15,8 +15,15 @@
   $json = file_get_contents('php://input');
   $data = json_decode($json, true);
 
-  // $violation->id = isset($data['id']) ? $data['id'] : null;
-  $violation->id = isset($_GET['id']) ? $_GET['id'] : null;
-  $read = $violation->deleteViolation();
-  var_dump($read);
+  session_start();
+  if(isset($_SESSION["admin"])) {
+    // $violation->id = isset($data['id']) ? $data['id'] : null;
+    $violation->id = isset($_GET['id']) ? $_GET['id'] : null;
+    $read = $violation->deleteViolation();
+    if ($read) {
+      echo true;
+    } else {
+      echo false;
+    }
+  }
 ?>
